Branch 11

+++ TMS.Web/Controllers/LabelController.cs
@@ -1,11 +1,17 @@
 ﻿using Microsoft.AspNetCore.Mvc;
 using TMS.EntitiesDTO;
 using TMS.Business.Services;
+using Microsoft.AspNet.Identity;
+using Microsoft.AspNetCore.Authorization;
+using TMS.Web.Models;
+using System.Diagnostics;
 
 namespace TMS.Web.Controllers
 {
+    [Authorize]
     public class LabelController : Controller
     {
+        private string _userId { get { return User.Identity.GetUserId(); } }
         private readonly LabelService _labelService;
         public LabelController(LabelService labelService)
         {
@@ -15,14 +21,20 @@ namespace TMS.Web.Controllers
         [HttpGet]
         public IActionResult List()
         {
-            var labels = _labelService.GetAll();
+            var labels = _labelService.GetAll(_userId);
             return View(labels);
         }
 
         [HttpGet]
         public IActionResult Create()
         {
-            return View(new LabelDTO());
+            if(_userId==null)
+                return View(new ErrorViewModel { RequestId = Activity.Current?.Id ?? HttpContext.TraceIdentifier });
+            var label = new LabelDTO()
+            {
+               UserId= _userId
+            };
+            return View(label);
         }
 
         [HttpPost]
@@ -37,6 +49,10 @@ namespace TMS.Web.Controllers
         public IActionResult Edit(int id)
         {
             var label = _labelService.GetById(id);
+            if(label.UserId !=_userId) // when u try to get access to other people label
+            {
+                return View(new ErrorViewModel { RequestId = "U can`t get access to this label" });
+            }
             return View(label);
         }
 
@@ -52,9 +68,13 @@ namespace TMS.Web.Controllers
         [ValidateAntiForgeryToken]
         public IActionResult Delete(int id)
         {
+            var label = _labelService.GetById(id);
+            if (label.UserId != _userId) // when u try to delete other people label
+            {
+                return View(new ErrorViewModel { RequestId = "U can`t get access to this label" });
+            }
             _labelService.Delete(id);
-            var labels = _labelService.GetAll();
-            return View("List", labels);
+            return RedirectToAction(nameof(List));
         }
     }
 }
